[HOME_ACTIVITY] Filme do dia implementado na tela home.

+++ app/src/main/java/br/com/digitalhouse/playmovieapp/domain/Entities.kt
@@ -20,6 +20,6 @@ data class Result(
     val release_date: String,
     val title: String,
     val video: Boolean,
-    val vote_average: Int,
+    val vote_average: Double,
     val vote_count: Int
 )
